refactor: flatten boot sequence into init/main.c style (orchestrator.py)
Rewrite create_nexus_fs() as a flat, top-to-bottom sequence — the
"init/main.c" of Nexus. Previously the boot sequence bounced between
orchestrator.py → _lifecycle.py → _wired.py → back to orchestrator.py
via callback injection (_link_fn/_initialize_fn). Now every step is
visible in one place.

- Inline _do_link() and _do_initialize() directly into create_nexus_fs()
- Extract _register_vfs_hooks() into new _hooks.py module
- Delete _lifecycle.py (dead code — no other callers)
- Set _linked/_initialized flags directly instead of injecting callbacks
